feat(editor): Provide workflow ID through composables instead of direct store access
Refactor components to use useProvideWorkflowId and useInjectWorkflowId composables
for accessing the workflow ID. This improves testability and reduces direct dependencies
on the workflows store.
