+++ packages/frontend/editor-ui/src/app/components/app/AppChatPanel.vue
@@ -2,6 +2,7 @@
 import AssistantsHub from '@/features/ai/assistant/components/AssistantsHub.vue';
 import { useChatPanelStore } from '@/features/ai/assistant/chatPanel.store';
 import { useUIStore } from '@/app/stores/ui.store';
+import { useProvideWorkflowId } from '@/app/composables/useProvideWorkflowId';
 import { computed, nextTick, onBeforeUnmount, onMounted, watch } from 'vue';
 
 const props = defineProps<{
@@ -10,6 +11,8 @@ const props = defineProps<{
 
 const chatPanelStore = useChatPanelStore();
 const uiStore = useUIStore();
+
+useProvideWorkflowId();
 
 const chatPanelWidth = computed(() => chatPanelStore.width);
