+++ packages/frontend/editor-ui/src/app/layouts/DemoLayout.vue
@@ -1,18 +1,9 @@
 <script lang="ts" setup>
-import { provide, computed } from 'vue';
-import { useRoute } from 'vue-router';
 import BaseLayout from './BaseLayout.vue';
 import DemoFooter from '@/features/execution/logs/components/DemoFooter.vue';
-import { WorkflowIdKey } from '@/app/constants/injectionKeys';
+import { useProvideWorkflowId } from '@/app/composables/useProvideWorkflowId';
 
-const route = useRoute();
-
-const workflowId = computed(() => {
-	const name = route.params.name;
-	return (Array.isArray(name) ? name[0] : name) as string;
-});
-
-provide(WorkflowIdKey, workflowId);
+useProvideWorkflowId();
 </script>
 
 <template>
