+++ packages/frontend/editor-ui/src/app/layouts/WorkflowLayout.vue
@@ -1,25 +1,17 @@
 <script lang="ts" setup>
-import { provide, computed } from 'vue';
-import { useRoute } from 'vue-router';
 import BaseLayout from './BaseLayout.vue';
 import { useLayoutProps } from '@/app/composables/useLayoutProps';
+import { useProvideWorkflowId } from '@/app/composables/useProvideWorkflowId';
 import AskAssistantFloatingButton from '@/features/ai/assistant/components/Chat/AskAssistantFloatingButton.vue';
 import { useAssistantStore } from '@/features/ai/assistant/assistant.store';
 import AppHeader from '@/app/components/app/AppHeader.vue';
 import AppSidebar from '@/app/components/app/AppSidebar.vue';
 import LogsPanel from '@/features/execution/logs/components/LogsPanel.vue';
-import { WorkflowIdKey } from '@/app/constants/injectionKeys';
 
-const route = useRoute();
 const { layoutProps } = useLayoutProps();
 const assistantStore = useAssistantStore();
 
-const workflowId = computed(() => {
-	const name = route.params.name;
-	return (Array.isArray(name) ? name[0] : name) as string;
-});
-
-provide(WorkflowIdKey, workflowId);
+useProvideWorkflowId();
 </script>
 
 <template>
