+++ packages/frontend/editor-ui/src/features/ndv/runData/components/RunData.vue
@@ -49,6 +49,7 @@ import { useNodeType } from '@/app/composables/useNodeType';
 import type { PinDataSource, UnpinDataSource } from '@/app/composables/usePinnedData';
 import { usePinnedData } from '@/app/composables/usePinnedData';
 import { useTelemetry } from '@/app/composables/useTelemetry';
+import { useInjectWorkflowId } from '@/app/composables/useInjectWorkflowId';
 import { useToast } from '@/app/composables/useToast';
 import { dataPinningEventBus } from '@/app/event-bus';
 import { ndvEventBus } from '@/features/ndv/shared/ndv.eventBus';
@@ -233,6 +234,7 @@ const schemaPreviewStore = useSchemaPreviewStore();
 
 const toast = useToast();
 const route = useRoute();
+const workflowId = useInjectWorkflowId();
 const nodeHelpers = useNodeHelpers();
 const externalHooks = useExternalHooks();
 const telemetry = useTelemetry();
@@ -880,7 +882,7 @@ function onItemHover(itemIndex: number | null) {
 
 function onClickDataPinningDocsLink() {
 	telemetry.track('User clicked ndv link', {
-		workflow_id: workflowsStore.workflowId,
+		workflow_id: workflowId.value,
 		push_ref: props.pushRef,
 		node_type: activeNode.value?.type,
 		pane: 'output',
@@ -956,7 +958,7 @@ function enterEditMode({ origin }: EnterEditModeArgs) {
 		is_output_present: hasNodeRun.value || pinnedData.hasData.value,
 		view: !hasNodeRun.value && !pinnedData.hasData.value ? 'undefined' : props.displayMode,
 		is_data_pinned: pinnedData.hasData.value,
-		workflow_id: workflowsStore.workflowId,
+		workflow_id: workflowId.value,
 		node_id: activeNode.value?.id,
 	});
 }
@@ -1033,7 +1035,7 @@ async function onTogglePinData({ source }: { source: PinDataSource | UnpinDataSo
 			push_ref: props.pushRef,
 			run_index: props.runIndex,
 			view: !hasNodeRun.value && !pinnedData.hasData.value ? 'none' : props.displayMode,
-			workflow_id: workflowsStore.workflowId,
+			workflow_id: workflowId.value,
 			node_id: activeNode.value?.id,
 		};
 
@@ -1093,7 +1095,7 @@ function showTooMuchData() {
 	userEnabledShowData.value = true;
 	telemetry.track('User clicked ndv button', {
 		node_type: activeNode.value?.type,
-		workflow_id: workflowsStore.workflowId,
+		workflow_id: workflowId.value,
 		push_ref: props.pushRef,
 		pane: props.paneType,
 		type: 'showTooMuchData',
@@ -1120,7 +1122,7 @@ function onCurrentPageChange(value: number) {
 	currentPage.value = value;
 	telemetry.track('User changed ndv page', {
 		node_type: activeNode.value?.type,
-		workflow_id: workflowsStore.workflowId,
+		workflow_id: workflowId.value,
 		push_ref: props.pushRef,
 		pane: props.paneType,
 		page_selected: currentPage.value,
@@ -1143,7 +1145,7 @@ function onPageSizeChange(newPageSize: number) {
 
 	telemetry.track('User changed ndv page size', {
 		node_type: activeNode.value?.type,
-		workflow_id: workflowsStore.workflowId,
+		workflow_id: workflowId.value,
 		push_ref: props.pushRef,
 		pane: props.paneType,
 		page_selected: currentPage.value,
@@ -1176,7 +1178,7 @@ function onDisplayModeChange(newDisplayMode: IRunDataDisplayMode) {
 			previous_view: previous,
 			new_view: newDisplayMode,
 			node_type: activeNode.value.type,
-			workflow_id: workflowsStore.workflowId,
+			workflow_id: workflowId.value,
 			push_ref: props.pushRef,
 			pane: props.paneType,
 		});
